// src/app/pages/employer/models/icc.ts
export interface Icc{
    name: String;
  mobileNumber: String;
  emailId: String;
  designation: String;
  department: String;
  experience: String;
  dateOfNomination: String;
  gender: String;
  organisationName: String;
  regNumber: String;
  unitName: String;
  userCategory: String;
  userSubCategory: String
}